PS: Implement constraint fixes for cartesian direction
Basically the constraint model is a bit of a bodge, so line and plane constraints cause instability for very stiff problems. I added a fix, but this implementation works only if the constraints are oriented along the x, y, or z axis. Any deviation and the constraint remains in its bodged state.

It is possible to fix it, but it would require some  change of variable to the appropriate subspace (resulting DOF) and to propagate that to the jacobian matrices.